requirements.txt
Django==1.9.8
django-cors-headers==1.1.0
djangorestframework==3.4.0
PyMySQL==0.7.5
requests==2.10.0
